Token refresh WIP, refactoring authentication process
