+++ src/Layout/Header/Header.js
@@ -4,7 +4,6 @@ import { useIntl } from 'react-intl';
 /** Services */
 import http from '@services/http';
 import { login } from '@services/auth';
-import { disableRepeat } from '@services/spotify';
 /** Actions */
 import { setTokenAction, logoutAction } from '@context/auth';
 import { fetchMeAction } from '@context/me';
@@ -15,6 +14,7 @@ import Button from '@common/Button/Button';
 import Select from '@common/Select';
 import { Popover, PopoverTrigger } from '@common/Popover';
 import UserAvatar from '@layout/Header/UserAvatar';
+import { retrieveSpotifyToken } from '@utils';
 import DeviceSelector from './DeviceSelector';
 /** Styled components */
 import {
@@ -45,36 +45,16 @@ const Header = () => {
   };
 
   useEffect(() => {
-    if (token) {
-      http.setToken(token);
-    } else {
-      const hash = window.location.hash
-        .substring(1)
-        .split('&')
-        .reduce((initial, item) => {
-          if (item) {
-            const parts = item.split('=');
-            initial[parts[0]] = decodeURIComponent(parts[1]);
-          }
-          return initial;
-        }, {});
-
-      const _token = hash.access_token;
-
-      window.location.hash = '';
-
+    if (window.location.hash) {
+      const _token = retrieveSpotifyToken();
       _token && dispatch(setTokenAction(_token));
     }
   }, []);
 
   useEffect(() => {
-    async function disableRepeatFn() {
-      await disableRepeat();
-    }
     if (token) {
       http.setToken(token);
       dispatch(fetchDevicesAction());
-      disableRepeatFn();
       dispatch(fetchMeAction());
     }
   }, [token]);
@@ -83,6 +63,14 @@ const Header = () => {
     <HeaderWrapper>
       <HeaderLogo>Rogue Sound</HeaderLogo>
       <HeaderActionsWrapper>
+        <button
+          onClick={() =>
+            dispatch(setTokenAction('testing-token-renewal-sorry-spotify'))
+          }
+          type="button"
+        >
+          Fuck up my token (Testing)
+        </button>
         {token && <DeviceSelector intl={intl} />}
         <HeaderLanguage>
           <Select
